refactor vade plugin handling
- remove `DidResolver`, `VcResolver`, `Logger`, `MessageHandler` traits in favour of `VadePlugin`
- split lib and traits file into `vade` and `vade_plugins` files, import those in `lib`
- all exports are now top-level (`Vade`, `VadePlugin`, `VadePluginResultValue`)
- remove `StorageCache` example plugin
- [CORE-1401]
